# backend/src/main/java/com/xrm/generated/api/XRmApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.xrm.generated.api;

import com.xrm.generated.model.CustomerDTO;
import com.xrm.generated.model.ReportDTO;
import com.xrm.generated.model.ResponsibleDTO;
import com.xrm.generated.model.TaskDTO;
import com.xrm.generated.model.TaskDetailsDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-09T19:31:32.650574400+01:00[Europe/Budapest]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "xRM", description = "the xRM API")
public interface XRmApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /customers/{customerId} : Delete customer
     *
     * @param customerId ID of the customer (required)
     * @return description (status code 200)
     */
    @Operation(
        operationId = "deleteCustomer",
        summary = "Delete customer",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "description")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/customers/{customerId}"
    )
    
    default ResponseEntity<Void> deleteCustomer(
        @Parameter(name = "customerId", description = "ID of the customer", required = true, in = ParameterIn.PATH) @PathVariable("customerId") Integer customerId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /task/{taskId} : Delete task
     *
     * @param taskId ID of the task (required)
     * @return description (status code 200)
     */
    @Operation(
        operationId = "deleteTask",
        summary = "Delete task",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "description")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/task/{taskId}"
    )
    
    default ResponseEntity<Void> deleteTask(
        @Parameter(name = "taskId", description = "ID of the task", required = true, in = ParameterIn.PATH) @PathVariable("taskId") Integer taskId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /customers : Get all customer.
     *
     * @return List of customers (status code 200)
     */
    @Operation(
        operationId = "getAllCustomer",
        summary = "Get all customer.",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of customers", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CustomerDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customers",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<CustomerDTO>> getAllCustomer(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /responsible : Get all responsible user
     *
     * @return List of responsible user (status code 200)
     */
    @Operation(
        operationId = "getAllResponsibleUser",
        summary = "Get all responsible user",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of responsible user", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ResponsibleDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/responsible",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<ResponsibleDTO>> getAllResponsibleUser(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /task : Get all Task.
     *
     * @return List of task (status code 200)
     */
    @Operation(
        operationId = "getAllTask",
        summary = "Get all Task.",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of task", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TaskDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/task",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<TaskDTO>> getAllTask(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /customers/{customerId} : Get customer tasks.
     *
     * @param customerId ID of the customer (required)
     * @return description (status code 200)
     */
    @Operation(
        operationId = "getCustomerTasks",
        summary = "Get customer tasks.",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "description", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TaskDetailsDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customers/{customerId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<TaskDetailsDTO>> getCustomerTasks(
        @Parameter(name = "customerId", description = "ID of the customer", required = true, in = ParameterIn.PATH) @PathVariable("customerId") Integer customerId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /report : get report
     *
     * @return Report (status code 200)
     */
    @Operation(
        operationId = "getReport",
        summary = "get report",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Report", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ReportDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/report",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<ReportDTO>> getReport(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /customers : Create customer
     *
     * @param customerDTO description (required)
     * @return description (status code 200)
     *         or description (status code 500)
     */
    @Operation(
        operationId = "newCustomer",
        summary = "Create customer",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "description"),
            @ApiResponse(responseCode = "500", description = "description")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/customers",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> newCustomer(
        @Parameter(name = "CustomerDTO", description = "description", required = true) @Valid @RequestBody CustomerDTO customerDTO
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /task : Create task
     *
     * @param taskDTO description (required)
     * @return description (status code 200)
     *         or description (status code 500)
     */
    @Operation(
        operationId = "newTask",
        summary = "Create task",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "description"),
            @ApiResponse(responseCode = "500", description = "description")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/task",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> newTask(
        @Parameter(name = "TaskDTO", description = "description", required = true) @Valid @RequestBody TaskDTO taskDTO
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /customers : Update customer
     *
     * @param customerDTO description (required)
     * @return description (status code 200)
     */
    @Operation(
        operationId = "updateCustomer",
        summary = "Update customer",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "description", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/customers",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<CustomerDTO> updateCustomer(
        @Parameter(name = "CustomerDTO", description = "description", required = true) @Valid @RequestBody CustomerDTO customerDTO
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /task : Update task
     *
     * @param taskDTO description (required)
     * @return description (status code 200)
     */
    @Operation(
        operationId = "updateTask",
        summary = "Update task",
        tags = { "xRM" },
        responses = {
            @ApiResponse(responseCode = "200", description = "description", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/task",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<TaskDTO> updateTask(
        @Parameter(name = "TaskDTO", description = "description", required = true) @Valid @RequestBody TaskDTO taskDTO
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}